fix: fix type hints for d_at

+++ razel/razel.py
@@ -8,7 +8,7 @@ from utils.rotations import rot2, rot3
 
 
 # TODO add missing argument type hints once types known.
-def razel(r_eci: np.ndarray, v_eci: np.ndarray, date_time: datetime, d_ut1: float, d_at: float, x_p: float, y_p: float, phi_gd: float, lda: float, h_ellp: float) -> list[float]:
+def razel(r_eci: np.ndarray, v_eci: np.ndarray, date_time: datetime, d_ut1: float, d_at: int, x_p: float, y_p: float, phi_gd: float, lda: float, h_ellp: float) -> list[float]:
     """Calculates a spacecraft's range, azimuth, elevation, and the rates of these, from an observation site on the ground.
 
     :param r_eci: spacecraft position vector in the Earth-centred inertial (ECI) reference frame.
@@ -108,7 +108,7 @@ def razel(r_eci: np.ndarray, v_eci: np.ndarray, date_time: datetime, d_ut1: floa
 
 
 # TODO add missing argument type hints once types known.
-def fk5(r_eci: np.ndarray, v_eci: np.ndarray, date_time: datetime, d_ut1: float, d_at: float, x_p: float, y_p: float) -> list[np.ndarray]:
+def fk5(r_eci: np.ndarray, v_eci: np.ndarray, date_time: datetime, d_ut1: float, d_at: int, x_p: float, y_p: float) -> list[np.ndarray]:
     """
     Converts position and velocity vectors in the ECI frame to the ECEF frame.
